Create a database file , and add some constraints for Request validation

+++ REST-API-Validation-Exception-Handling/src/main/java/com/RomayssaeAPI/Validation/Exception/Handling/service/UserService.java
@@ -29,7 +29,7 @@ public class UserService {
         return repository.findAll();
       }
 
-      public User getUser(Long id){
+    public User getUser(int id){
         return repository.findByUserId(id);
-      }
+    }
 }
